+++ Net_Core/Framework/Enumerations/FunctionSecurityTableNames.cs
@@ -2,6 +2,7 @@ namespace GrowthWare.Framework.Enumerations;
 
 public class FunctionSecurityTableNames
 {
+    public const string FUNCTIONS = "Functions";
     public const string DERIVED_ROLES = "DerivedRoles";
     public const string ASSIGNED_ROLES = "AssignedRoles";
     public const string ASSIGNED_GROUPS = "Groups";