+++ Net_Core/Framework/Enumerations/FunctionSecurityTables.cs
@@ -5,4 +5,5 @@ public enum FunctionSecurityTables
     DerivedRoles = 0,
     AssignedRoles = 1,
     AssignedGroups = 2,
+    Functions = 3,
 }